Add isCommanderArmy toggle to Spectator HUD
This adds a togglable flag to Spectator HUD. If toggled false, Commanders will not be counted for in Army value. This brings a small QoL benefit for those who wish to track small army values at the beginning of games, especially for 1v1s where there can be only a single Commander per player.